feat: remove `diabot nightscoutgraph` and `diabot nightscoutgraph mode` commands.
text-based commands should be avoided in favor of slash-commands given discord's push to remove message content permissions from bots, so these text commands should be removed before being pushed to production
