+++ bot/src/main/java/com/dongtronic/diabot/Main.kt
@@ -12,7 +12,10 @@ import com.dongtronic.diabot.platforms.discord.commands.diabetes.EstimationComma
 import com.dongtronic.diabot.platforms.discord.commands.info.AboutCommand
 import com.dongtronic.diabot.platforms.discord.commands.info.InfoCommand
 import com.dongtronic.diabot.platforms.discord.commands.misc.*
-import com.dongtronic.diabot.platforms.discord.commands.nightscout.*
+import com.dongtronic.diabot.platforms.discord.commands.nightscout.NightscoutAdminCommand
+import com.dongtronic.diabot.platforms.discord.commands.nightscout.NightscoutApplicationCommand
+import com.dongtronic.diabot.platforms.discord.commands.nightscout.NightscoutCommand
+import com.dongtronic.diabot.platforms.discord.commands.nightscout.NightscoutGraphApplicationCommand
 import com.dongtronic.diabot.platforms.discord.commands.quote.QuoteCommand
 import com.dongtronic.diabot.platforms.discord.commands.rewards.RewardsCommand
 import com.dongtronic.diabot.platforms.discord.listeners.*
@@ -88,7 +91,6 @@ object Main {
                 // BG
                 ConvertCommand(bgCategory),
                 NightscoutCommand(bgCategory),
-                NightscoutGraphCommand(bgCategory),
 
                 // Utility
                 PingCommand(utilitiesCategory),
