+++ bot/src/main/java/com/dongtronic/diabot/platforms/discord/commands/nightscout/NightscoutGraphApplicationCommand.kt
@@ -1,9 +1,17 @@
 package com.dongtronic.diabot.platforms.discord.commands.nightscout
 
 import com.dongtronic.diabot.data.mongodb.GraphDisableDAO
+import com.dongtronic.diabot.data.mongodb.NightscoutDAO
 import com.dongtronic.diabot.exceptions.NightscoutFetchException
+import com.dongtronic.diabot.exceptions.UnconfiguredNightscoutException
+import com.dongtronic.diabot.graph.BgGraph
 import com.dongtronic.diabot.platforms.discord.commands.ApplicationCommand
 import com.dongtronic.diabot.util.logger
+import com.dongtronic.nightscout.EntriesParameters
+import com.dongtronic.nightscout.Nightscout
+import com.dongtronic.nightscout.data.NightscoutDTO
+import com.dongtronic.nightscout.exceptions.NoNightscoutDataException
+import com.fasterxml.jackson.core.JsonProcessingException
 import dev.minn.jda.ktx.await
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.reactor.awaitSingle
@@ -12,8 +20,15 @@ import net.dv8tion.jda.api.events.interaction.ButtonClickEvent
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent
 import net.dv8tion.jda.api.interactions.commands.build.CommandData
 import org.knowm.xchart.BitmapEncoder
+import org.knowm.xchart.XYChart
+import org.litote.kmongo.MongoOperator
+import reactor.core.publisher.Mono
+import retrofit2.HttpException
+import java.net.UnknownHostException
 import java.time.Duration
+import java.time.Instant
 import kotlin.math.abs
+import kotlin.math.max
 
 class NightscoutGraphApplicationCommand : ApplicationCommand {
     override val commandName: String = "graph"
@@ -46,7 +61,7 @@ class NightscoutGraphApplicationCommand : ApplicationCommand {
                         return@launch
                     }
 
-                    val chart = NightscoutGraphCommand.getDataSet(event.user.id).awaitSingle()
+                    val chart = getDataSet(event.user.id).awaitSingle()
                     val imageBytes = BitmapEncoder.getBitmapBytes(chart, BitmapEncoder.BitmapFormat.PNG)
                     hook.await().editOriginal(imageBytes, "graph.png").submit().await()
                     applyCooldown(event.user.id)
@@ -89,5 +104,44 @@ class NightscoutGraphApplicationCommand : ApplicationCommand {
 
     override fun config(): CommandData {
         return CommandData(commandName, "Generate a graph from Nightscout")
+    }
+
+    private fun getDataSet(senderId: String, time: Duration = Duration.ofHours(4)): Mono<XYChart> {
+        return NightscoutDAO.instance.getUser(senderId)
+                .zipWhen { userDTO ->
+                    if (userDTO.url == null) {
+                        return@zipWhen Mono.error<NightscoutDTO>(UnconfiguredNightscoutException())
+                    }
+
+                    val ns = Nightscout(userDTO.url, userDTO.token)
+
+                    // calculate the amount of readings there should be.
+                    // assume 1 reading every 5 minutes and a minimum of 1 reading
+                    val count = max(time.toMinutes() / 5, 1).toInt()
+                    val startTime = Instant.now()
+                            .minus(time)
+                            .toEpochMilli()
+                            .toString()
+                    val findParam = EntriesParameters()
+                            .find("sgv", operator = MongoOperator.exists)
+                            .find("date", startTime, MongoOperator.gte)
+                            .count(count)
+                            .toMap()
+                    ns.getSgv(params = findParam)
+                            // duplicate code from NightscoutCommand. will be cleaned up later with a refactor of both
+                            .onErrorMap({ error ->
+                                error is HttpException
+                                        || error is UnknownHostException
+                                        || error is JsonProcessingException
+                                        || error is NoNightscoutDataException
+                            }, {
+                                NightscoutFetchException(userDTO, it)
+                            })
+                            .flatMap { ns.getSettings(it) }
+                }.map { tuple ->
+                    val userDTO = tuple.t1
+                    val ns = tuple.t2
+                    BgGraph(userDTO.graphSettings).addEntries(ns)
+                }
     }
 }
